server/requirements.txt
click==7.1.2
Flask==1.1.2
gevent==21.1.2
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-pytun==2.4.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0